django/requirements.txt
Django==3.0.7
django-markdownx